[🚚rename]: deliveryNote -> delivery 로 파일명 단순화
